+++ src/components/core/compute/getResults.ts
@@ -2,7 +2,6 @@ import { P2PCommandResponse } from '../../../@types/index.js'
 import { CORE_LOGGER } from '../../../utils/logging/common.js'
 import { Handler } from '../handler/handler.js'
 import { ComputeGetResultCommand } from '../../../@types/commands.js'
-import { C2DEngine } from '../../c2d/compute_engines.js'
 import { checkNonce, NonceResponse } from '../utils/nonceHandler.js'
 import {
   buildInvalidRequestMessage,
@@ -75,7 +74,7 @@ export class ComputeGetResultHandler extends Handler {
     // env might contain
     let engine
     try {
-      engine = await C2DEngine.getC2DByHash(hash)
+      engine = await this.getOceanNode().getC2DEngines().getC2DByHash(hash)
     } catch (e) {
       return {
         stream: null,
